Modified vector product fifo
so that full and empty signals are no longer latched and the module now
becomes a thin wrapper around the standard fifo. This is done because so
long as the fifo size is correct, the standard fifo is safe to use as
the intermediate result store between the forward prop producer and
backprop consumer, and the extra logic is unneeded.

if there turns out to be problems down the line (unlikely), we can
    always revert